# openapi/nadir-curated.yaml
...
openapi: 3.0.3
info:
  title: Nadir Gateway (curated contract)
  version: 0.1.0
  description: >-
    Subset of the Nadir gateway API guaranteed to remain OpenAI-compatible
    on request/response shapes. Lifecycle (wake/idle) is tested separately (pytest -m smoke).
paths:
...
  /v1/audio/transcriptions:
    post:
      operationId: createTranscription
      summary: Speech-to-text (multipart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AudioTranscriptionRequest"
      responses:
        "200":
          description: Transcription JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "503":
          $ref: "#/components/responses/UnavailableError"
  /v1/audio/transcriptions/stream:
    post:
      operationId: createTranscriptionStream
      summary: Speech-to-text streaming (SSE)
      description: |
        Multipart upload with partial transcript events over Server-Sent Events.
        Not OpenAI Realtime WebSocket — custom `transcript` and `done` event types.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AudioTranscriptionStreamRequest"
      responses:
        "200":
          description: SSE transcript stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Frames: `event: transcript` with JSON delta/completed payloads,
                  then `event: done`.
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "503":
          $ref: "#/components/responses/UnavailableError"
  /v1/audio/translations:
    post:
      operationId: createTranslation
      summary: Speech translation (multipart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AudioTranscriptionRequest"
      responses:
        "200":
          description: Translation JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscriptionResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "503":
          $ref: "#/components/responses/UnavailableError"
components:
  responses:
    BadRequestError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenAIErrorResponse"
    NotFoundError:
      description: Model not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenAIErrorResponse"
    UnavailableError:
      description: Model unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OpenAIErrorResponse"
  schemas:
...
    AudioTranscriptionRequest:
      type: object
      required: [model, file]
      properties:
        model:
          type: string
        file:
          type: string
          format: binary
        response_format:
          type: string
      additionalProperties: true
    AudioTranscriptionStreamRequest:
      type: object
      required: [model, file]
      properties:
        model:
          type: string
        file:
          type: string
          format: binary
        language:
          type: string
        chunk_duration:
          type: number
        word_timestamps:
          type: boolean
        prompt:
          type: string
        temperature:
          type: number
      additionalProperties: true
    TranscriptionResponse:
      type: object
      required: [text]
      properties:
        text:
          type: string
      additionalProperties: true
...
    OpenAIErrorResponse:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message:
              type: string
            type:
              type: string
            code:
              type: string
          additionalProperties: true